requirements.txt
alembic==1.8.1
bcrypt==4.0.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
packaging==21.3
Pillow==9.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
